requirements.txt
annotated-types==0.7.0
anyio==4.4.0
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.112.1
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
pillow==10.4.0
psycopg2==2.9.9
PyAutoGUI==0.9.54
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
python-telegram-bot==21.4
pytweening==1.2.0
pytz==2024.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.4
wikipedia==1.4.0
